refactor: 💡 change request names
